refactor(bench): factor Mooncake utilities out of Claude exporter
Split lib/bench/src/coding/mooncake.rs into a producer-agnostic primitives
layer so future bench/export modules (e.g. a Dynamo agent trace exporter)
can emit Mooncake replay traces without depending on Claude parser types.

mooncake.rs (generic):
- MooncakeRow with optional timestamp/delay (skipped on serialize when None)
- RollingHashIdMapper for sequence-aware block-hash-to-id mapping, reusable
  across requests so equal blocks across requests get equal ids
- hash_token_blocks helper over &[u32] using dynamo_tokens::compute_hash_v2
- MooncakeJsonlWriter for row + optional sidecar JSONL streams, paired with
  the existing sidecar_path_for convention in common.rs
- write_empty_files for the no-rows-produced path

claude/export.rs (Claude orchestration):
- ExportConfig, ExportStats, write_streamed_mooncake_rows
- SessionState, HeadTurn, OverlapBase, ReadyTurn, HeapEntry, TokenizeJob
- Tokenizer worker pool, scheduling, global ordering across sessions
- Consumes the generic mooncake primitives to emit rows + sidecar

No behavior changes: same JSON schema, same CLI surface, same parsing
semantics. Hash ids remain workload identity labels for replay, not
literal Dynamo runtime KV hashes.

Tests cover: shared prefix produces shared leading hash_ids, mapper state
preserved across requests, optional row fields skipped on serialize,
sidecar path convention preserved, writer rejects sidecar writes when not
configured.
