+++ lib/bench/src/coding/claude/mod.rs
@@ -2,4 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod discovery;
+pub mod export;
 pub mod parser;
